Moved observations & training data
